Crate sui_macros

Macros

checked_arithmetic

fail_point

fail_point_arg

fail_point_async

fail_point_if

nondeterministic

Simply evaluates expr.

replay_log

Use to write INFO level logs only when REPLAY_LOG environment variable is set. Useful for log lines that are only relevant to test infra which still may need to run a release build. Also note that since logs of a chain replay are exceedingly verbose, this will allow one to bubble up “debug level” info while running with RUST_LOG=info.

Statics

ANTITHESIS_ASSERTIONS_ENABLED

Functions

clear_fail_point

handle_fail_point

handle_fail_point_arg

handle_fail_point_async

handle_fail_point_if

register_fail_point

register_fail_point_arg

Register code to run locally if the fail point is hit, with a value provided by the test. If the registered callback returns a Some(v), then the v is passed to the callback in the test.

register_fail_point_async

Register an asynchronous fail point. Because it is async it can yield execution of the calling task, e.g. by sleeping.

register_fail_point_if

Register code to run locally if the fail point is hit. Example:

register_fail_points

Attribute Macros

init_static_initializers

sim_test

The sim_test macro will invoke #[msim::test] if the simulator config var is enabled.

sui_test

The sui_test macro will invoke either #[msim::test] or #[tokio::test], depending on whether the simulator config var is enabled.

with_checked_arithmetic

Derive Macros

EnumVariantOrder

This proc macro generates a function order_to_variant_map which returns a map of the position of each variant to the name of the variant. It is intended to catch changes in enum order when backward compat is required.