Function sui_macros::register_fail_point_arg

pub fn register_fail_point_arg<T: Send + 'static>(
    identifier: &'static str,
    callback: impl Fn() -> Option<T> + Sync + Send + 'static,
)

Register code to run locally if the fail point is hit, with a value provided by the test. If the registered callback returns a Some(v), then the v is passed to the callback in the test.

In the test:

    register_fail_point_arg("foo", || {
        Some(42)
    });

In the code:

    let mut value = 0;
    fail_point_arg!("foo", |arg| {
       value = arg;
    });