sui_macros
Macro replay_log

macro_rules! replay_log {
    ($($arg:tt)+) => { ... };
}

Use to write INFO level logs only when REPLAY_LOG environment variable is set. Useful for log lines that are only relevant to test infra which still may need to run a release build. Also note that since logs of a chain replay are exceedingly verbose, this will allow one to bubble up “debug level” info while running with RUST_LOG=info.